+++ react/components/icon/Success/index.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import config from 'vtex-tachyons/config.json'
-import { calcIconSize } from '../utils'
+import { calcIconSize, baseClassname } from '../utils'
 
 const iconBase = {
   width: 16,
@@ -16,6 +16,7 @@ class Success extends PureComponent {
     if (solid) {
       return (
         <svg
+          className={baseClassname('upload', 'solid')}
           width={newSize.width}
           height={newSize.height}
           viewBox="0 0 16 16"
@@ -31,6 +32,7 @@ class Success extends PureComponent {
     }
     return (
       <svg
+        className={baseClassname('upload')}
         width={newSize.width}
         height={newSize.height}
         viewBox="0 0 16 16"